backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
openai
python-dotenv